Move some code from QuickvaluesWidgetStrategy into base class
We will reuse this soon.
