+++ graylog2-server/src/main/java/org/graylog2/dashboards/widgets/strategies/QuickvaluesWidgetStrategy.java
@@ -16,7 +16,6 @@
  */
 package org.graylog2.dashboards.widgets.strategies;
 
-import com.google.common.base.Splitter;
 import com.google.common.collect.Maps;
 import com.google.inject.assistedinject.Assisted;
 import com.google.inject.assistedinject.AssistedInject;
@@ -28,66 +27,25 @@ import org.graylog2.plugin.dashboards.widgets.ComputationResult;
 import org.graylog2.plugin.dashboards.widgets.WidgetStrategy;
 import org.graylog2.plugin.indexer.searches.timeranges.TimeRange;
 
-import javax.annotation.Nullable;
-import java.util.List;
-import java.util.Locale;
 import java.util.Map;
 
 import static com.google.common.base.MoreObjects.firstNonNull;
 import static com.google.common.base.Strings.isNullOrEmpty;
 
-public class QuickvaluesWidgetStrategy implements WidgetStrategy {
+public class QuickvaluesWidgetStrategy extends QuickvaluesBaseWidgetStrategy {
 
     public interface Factory extends WidgetStrategy.Factory<QuickvaluesWidgetStrategy> {
         @Override
         QuickvaluesWidgetStrategy create(Map<String, Object> config, TimeRange timeRange, String widgetId);
     }
 
-    private final String query;
-    @Nullable
-    private final String streamId;
-
-    private final String field;
-    private final Searches searches;
-    private final TimeRange timeRange;
-    private final String sortOrder;
     private final int dataTableLimit;
-    private final List<String> stackedFields;
 
     @AssistedInject
     public QuickvaluesWidgetStrategy(Searches searches, @Assisted Map<String, Object> config, @Assisted TimeRange timeRange, @Assisted String widgetId) throws InvalidWidgetConfigurationException {
-        this.searches = searches;
-        this.timeRange = timeRange;
+        super(searches, timeRange, config, widgetId);
 
-        if (!checkConfig(config)) {
-            throw new InvalidWidgetConfigurationException("Missing or invalid widget configuration. Provided config was: " + config.toString());
-        }
-
-        this.query = (String)config.get("query");
-
-        this.field = (String) config.get("field");
-        this.streamId = (String) config.get("stream_id");
-
-        this.sortOrder = (String) firstNonNull(config.get("sort_order"), "desc");
         this.dataTableLimit = (int) firstNonNull(config.get("data_table_limit"), 50);
-        this.stackedFields = getStackedFields(config.get("stacked_fields"));
-    }
-
-    private static List<String> getStackedFields(@Nullable Object value) {
-        final String stackedFieldsString = (String) firstNonNull(value, "");
-        return Splitter.on(',').trimResults().omitEmptyStrings().splitToList(stackedFieldsString);
-    }
-
-    private static Sorting.Direction getSortingDirection(String sort) {
-        if (isNullOrEmpty(sort)) {
-            return Sorting.Direction.DESC;
-        }
-
-        try {
-            return Sorting.Direction.valueOf(sort.toUpperCase(Locale.ENGLISH));
-        } catch (Exception e) {
-            return Sorting.Direction.DESC;
-        }
     }
 
     @Override
@@ -107,9 +65,5 @@ public class QuickvaluesWidgetStrategy implements WidgetStrategy {
         result.put("missing", terms.getMissing());
 
         return new ComputationResult(result, terms.tookMs());
-    }
-
-    private boolean checkConfig(Map<String, Object> config) {
-        return config.containsKey("field");
     }
 }
